+++ infra/resources/_modules/cms_function_app/rbac.tf
@@ -1,6 +1,6 @@
-module "function_app_to_blob_data_writer" {
+module "cms_fn_roles" {
   source  = "pagopa-dx/azure-role-assignments/azurerm"
-  version = "~> 1.0"
+  version = "~> 1.2"
 
   principal_id    = module.cms_fn.function_app.function_app.principal_id
   subscription_id = data.azurerm_subscription.current.subscription_id
@@ -12,11 +12,17 @@ module "function_app_to_blob_data_writer" {
     container_name       = local.containers.activations.name
     }
   ]
+  apim = [{
+    name                = local.apim.name
+    resource_group_name = local.apim.resource_group_name
+    description         = "To allow Function App to work with APIM"
+    role                = "owner"
+  }]
 }
 
-module "function_app_staging_slot_to_blob_data_writer" {
+module "cms_fn_staging_slot_roles" {
   source          = "pagopa-dx/azure-role-assignments/azurerm"
-  version         = "~> 1.0"
+  version         = "~> 1.2"
   principal_id    = module.cms_fn.function_app.function_app.slot.principal_id
   subscription_id = data.azurerm_subscription.current.subscription_id
   storage_blob = [{
@@ -27,4 +33,10 @@ module "function_app_staging_slot_to_blob_data_writer" {
     container_name       = local.containers.activations.name
     }
   ]
+  apim = [{
+    name                = local.apim.name
+    resource_group_name = local.apim.resource_group_name
+    description         = "To allow Function App (staging slot) to work with APIM"
+    role                = "owner"
+  }]
 }