+++ packages/external-clients/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @io-services-cms/external-clients
 
+## 1.8.4
+
+### Patch Changes
+
+- 24b16966: bump version of @azure/identity to v.4.13.0
+
+## 1.8.3
+
+### Patch Changes
+
+- 9ea20575: bump version of azure library arm-apimanagement to v.10 in cms, backoffice and external-clients package projects
+
+## 1.8.2
+
+### Patch Changes
+
+- 38e3ae89: Use Managed Identity instead of Service Principal to interact with APIM
+
 ## 1.8.1
 
 ### Patch Changes